Use mixins to add functionality to test models
