+++ pytorch_lightning/testing/__init__.py
@@ -1,3 +1,3 @@
 from .lm_test_module import LightningTestModel
-from .no_val_end_module import NoValEndTestModel
-from .no_val_module import NoValModel
+from .lm_test_module_base import LightningTestModelBase
+from .lm_test_module_mixins import *
